requirements.txt
selenium>=4.6.0
pandas>=1.5.0
colorama>=0.4.0
python-dotenv>=1.0.0
